Fix Waveshare Epaper component always doing full-screen update on every boot.
Init the `full_update_every` counter (`at_update_`) to 1 instead of zero so that it doesn't flicker on every boot. This is best suited for low-power ESP32 modules that go to sleep between updates. It is also a more logical implementation of the `full_update_every` setting because it will count up to the set number before doing a full redraw. If a full date is desired on boot, then use public method `set_full_update_every(100)` after boot.